requirements.txt
google-genai
langchain
streamlit
python-dotenv
langchain
